Recover from send failures in smack's PacketWriter
To recover from send failures in smack's PacketWriter which would
normally cause the send message to be silently lossed, we now get noticed
and can queue the message for resending later.

For this, we need a special PacketListener in smack that notifies yaxim
that a message could not be sent. This is introduced in
https://github.com/tuxmobil/asmack/commit/5498d4b9a701219d095f27b4f79c17d9ac8c94b6.

For yaxim, the chat message DB was extended to store smack's packet ID for
each message so that messages can be identified when they could not be
sent. The already available mechanism for offline sending is then used to mark
them for later resending.