Rework ticker text on new message notifications
There are privacy concerns regarding the ticker text of incoming
messages. Ticker texts are readable even without unlocking the device.

So far we show the sender name, other apps also provide a preview on the
message content. Being consequent, we shouldn't show either if the user
prefers privacy over the preview.

This introduces a configuration options for this. If it's enables, which
is the default, both the sender's name and the first line of the message
is shown in the ticker.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>