Drop groups table
These days we only use the groups table to identify is some group was
added or removed. But this approach is not perfect either as it doesn't
allow us to find out if a group became virtually empty (no more online
members).

This commit just raises a change notification for the group URI along
with any content change. No relevant impact on the visualization of the
roster could be found so far. Instead, the code becomes much simpler.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>