Annotate methods calling Sherlock's invalidateOptionsMenu
Lint things we call the Honeycomb+ Activity method invalidateOptionsMenu
and stops the build while we are actually invoking the one always
provided by SherlockExpandableListActivity. Work around it by annotating
the methods that do this to be Honeycomb+ only - which is not true and
not nice, but the best we can do to keep Lint usable.

Signed-off-by: Jan Kiszka <jan.kiszka@web.de>