XEP-0092: Software Version implemented
Report app version in reply to version query IQ.

This will report the (translated) app_name string, the long version
number from build_version, including the date, and "Android" as the OS.

The latter is a violation of XEP-0092 security considerations, however
yaxim is only available on Android, and you can already identify yaxim
based on a `disco#info` query.

Maybe there will be a configuration option to disable that behavior one
day.