chat window: only bind/unbind if onPause/onResume also happened
The ChatWindow is using bindService to clean notifications as well as to
update google:queue status. onPause/onResume are not reliable because
these are called if a chat window is right behind the lockscreen.
onWindowFocusChanged is not reliable because it is called on popup
dialogs / context menu opening. Therefore, both are combined.