MUC: work around IRC transports incapable of preserving message IDs
When yaxim sends a multi-line message to an IRC transport, the message
is split into multiple individual one-liners. Unfortunately, some IRC
transports don't add your original message ID to any of the reflected
messages, so it is not possible to correlate the incoming reflection
with the original message.

The heuristics to work around that look like this:

 * I recently sent a message to the MUC, AND
 * the received message is from my participant JID, AND
   * the message ID matches OR
   * the message body matches OR
   * *NOW ADDED* the received message is the first line of the sent one.